mon/PGMap: fix client/recovery rate keeps too long for pools
the `pg_sum_old` will be assigned on every `mgr_tick_period` (default is 2s)
while the original `pg_pool_sum_old` will be only assigned if there are new
pg stats reported from OSD.

so the sample interval of pg_sum is `mgr_tick_period`, but pool stats may be
skipped to update, and the sample interval are multiples of `mgr_tick_period`.
and this will result in pool's client/recovery rate keeps too long.

this change taht split from `63df4060824316476e5e022e1a923c57090c2be7` in master
branch try to fix it.

+++ src/mon/PGMap.h
@@ -328,7 +328,7 @@ public:
   void update_pool_deltas(
     CephContext *cct,
     const utime_t ts,
-    const mempool::pgmap::unordered_map<uint64_t, pool_stat_t>& pg_pool_sum_old);
+    const mempool::pgmap::unordered_map<int32_t, pool_stat_t>& pg_pool_sum_old);
   void clear_delta();
 
   void deleted_pool(int64_t pool) {
